student-management-system/backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
mysql-connector-python==9.4.0
gunicorn==21.2.0
Werkzeug==3.0.3